infernosource-monorepo/apps/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
sqlmodel
asyncpg
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
whois
openai
redis
celery
httpx
beautifulsoup4
lxml
python-dotenv
scrapy
bs4 



